A better attempt to use inline Ajax driven Webform
This replaces current Drupal Dialog with an embeded, ajax driven
Webform and adds a new Widget that has no need for a calling button.
For Webform 8.x-5.2 the following in-review Patch needs to be applied:
reason, even when WerbformSubmission saving is disabled, Webform tries
to get an absolute canonical Link to the submission. That is not
possible since no saved entities lack a proper ->id