Toggle Node submit actions visibilty if moderation state or title is changed
This adds some extra JS to the humble attempt of declutter object ingest worklow by re-cluttering it again if someone needs to just change the title or the moderation state of an Object, basically showing node actions if that happens.
TODO: really document the use cases and why we are doing this. Decluttering per se can not be the only motivator here and it needs to respond to the actual worklow
TODO2: bring this into a JS setting per field widget, so people can deactivate this functionality if needed.