Prepare for consistent file and entity processing
@giancarlobi the idea here is to avoid Webform specificities when processing things into as:images, documents, application structures. This is not being used yet, but what this does now is it generates  new ap: element.

Example

 "ap:entitymapping": {
        "entity:file": [
            "images",
            "documents",
            "upload_associated_digital_images_secondary"
        ]
    },

Means those keys contain files and we can be certain it is so. (i mean we already do in this context, but i want it in any context)

How will this be used?
Instead of processing all files in the webform handler (or as we do know invoking the file persister service) we will just generate that structure and emit an json processing event with this new ap: key
Then a listener will take that structure and do all the same things we do here but in the Subscriber.

How this benefits us?
If you ingest a node via JSON API (means no webform) you can pass that structure into the JSON (a hint) and then the process of generating as:image, etc will be done in the same consistent way
https://www.drupal.org/project/webform/issues/3067958

Or the opossite: you pass an as:image, etc and if those point back to a field (dr:for) the ap:structure will be generated.

There will be some cross checking and some mangling but to be fair. The number of upload fields will always be limited.
Also, this way we can do the same for entity:node, entity:user, etc
Consistent, all the way from raw JSON to webform input, that is what i want and less, less, less code.