Deal with empty $form elements for Webform steps
Adds a simple check for the existance of a `#type` key on our webform alter hook. This part of the code just deals with adding diret link buttons to each webform step to allow people to go directly to a certain step and make corrections before Saving metadata, when on the final preview step. When this fails Webform data never gets persisted to the private storage used to keep values around.