Further Fixes for ISSUE-17
This was annoying but finally this should be fixed.

For some reason, When issuing a Display Mode Switch (during node edit/ingest) the ajax callback was validating the whole form, including the internals of the webform, forcing us (our fault of course) to delete the session information / private storage that kept data connected between Webform widget and the actual Webform multistep process. Result. If doing an ingest from scratch using the demo account which is heavy on ajax, and changing the display mode at the end, we ended with an empty JSON and a lot of shame. I had the bad luck to explore (explode!) this in public while showcasing Archipelago. This is new and is related to some internals of that switch field, still my fault. There are so many ways to avoid this while testing = ingest correctly,  but i always made the right ajax combinations to have this explode in my face!. Anyway. This brings some minor changes.
- First, we keep the initial widget state id consistent, without reassigning every time the form rebuilds. Not a big change but allows in a next pull to restore from private storage a draft (if you have to go an open the door for UPS and your computer ends without battery).
- Second, we don't explicitly empty the session private storage anymore. That is not needed, we let it expire, but that way we can also restore it.
- I also addded a humble attempt of #limit validation errors for the webform itself, so other elements in different trees don't affect this so much. Still can be improved.
- Finally, i added error logging in case private temp store fails (means fatal DB error, still good to alert people while black smoke comes out right?) during save but also during partial webform validation, to make sure that in the worst of the cases, people don't spend hours filling a form that will not be able to be saved anyway. I also dump the webform content to logs, so at least, for now, data is not lost and tbis can be further debugged by an admin.

/src/Plugin/Field/FieldWidget/StrawberryFieldWebFormWidget.php looks like it was changed a lot, but its mostly just wrong identation fixed, legacy of my first steps months ago. You can see without spaces when reviewing for the actual changes.

@giancarlobi second pair of eyes would be appreciated!