ISSUE-16: How many Malibus are in the world? Will they still exist in 30 years?
Adds Nominatim resolution to a quite complex working and silly-simplistic looking Webform Field. Has quite a lot of Ajax interaction and settings going here and there.  I extended an existing class and now i feel there is overhead on the parent which is not generic enough there. Will have to go back and implement some local methods to make this cleaner. But hey, it works!

Tested on single form elements and Multiple ones (slightly different logic).
Does not use all the n*3 elements we get via GEOJSON but a subset significat enough for metadata + the OSM id and type to get all again if needed.

@TODO: various so will keep this open. I would love to have the time to add a visual hint via leaflet map for the values we extract. Probably that needs to go into format_strawberryfield.
@TODO: better validation, what makes an entry like this valid?

  Also there are quite a few Malibus in the world, but for sanity we only return the first 5!

Lastly. Since this is about geography, world, places, environment, lets put a  context statement here. Coding too long/MUCHO server+ screens and too much DevOps hurts the environment for real and Devs rarely care. Not kidding, we are wasting too much energy as Developers and users miss-using valious resources for just the sake of it/ just to keep the ball rolling, profit, consumism and even ego. Not fun. FridaysforFuture. starting in 4 minutes