Moving forward. Reading Multiple Scenes now
WIP but will commit. Now, if the #default value is an array of scenes, it takes the first scene and pops it into the #default saving the rest of the scenes in a temporary form state.

I also remove the "Multiple" Option for this element, the idea is to have multiple Scenes in a Tour, but not Multiple Tours in a single Object. Too much overhead and adds little to the UX.

A lot of new overriden stuff but also some cleanup and less hardcoded options. Not ready, but way better than today in the morning. tomorrow we will have this done!