Gosh. Total refactor. Super de-factor
Still need to remove error_logs all over the place but now it actuall allows to change between Scenes, and add dynamically hotspots, swap back and forth and  only saves what we need, more and not less. Also works when starting from scratch. Well, looking good and not failing after so many tests...

Pretty tired, will just continue first hour in the morning. This is pretty sweet and i feel kinda the largest dev for realz in this panorama thingy we are into