Help Inline Rendered Webforms to get rid of the form element
Yes, sounds crazy. But right now we are rendering a form inside a form, and even when that is working, it is messing up with AJAX States, etc. This should keep the internal webform/form functionality, but since we already have a form element on the actual entity edit/add operation, we can simple, when outputting bring this into a Div! Ha. Hopefully this works. I know i will need more try/test before getting it right.