Give each button its own name. That solves Ajax conflict
Ajax was sometimes not firing correctly and the scene removal JS action command was targeting the wrong hotspot because of the reuse (on form rebuild after a delete or Scene change using the currently editing scene selector) of the hotspot button name and id's. Of course! Now it uses the actual unique hotspot ID which is formed by the scene number + the internal id. All works perfectly