Comment out the show save (for  now)
This is WIP and this JS will probably be refactored based on some settings passed. The previous behaviour of toggling NODE Save Button based on the workflow state was confusing and showed sometimes a save button that if pressed would do nothing other than reload the page. The reason for that is deeper but related to the fact that we have really multiple forms and the submit handler/validation handlers compete with each other. We have an ISSUE open for this (#26 i think?) and this will require constante improvements as we move Archipelago forward.