Adds a `@TODO`/ WE need to fix this templating issue!!
This is strange and makes no sense really. But hear me out!
We are flagging our Webform Element (the one that uses a Full Webform with overridden settings and displays inlide) uisng the #attributes key. This is set everytime the form renders, good and then used to suggest a theme template hint so we can get rid of a form inside a form HTML tag which is not valid HTML5. BUT! when the form does not validate this function is called, all is passed but! #attributes is not set! Means basically we end with the default template again and   form is rendered inside a form. Bad. This requires research and fixing. And of course sharper eyes.