WIP: Show edit button when the Node already exists (for metadata)
This is a good fix i feel, it just requires some settings.
If the node already exists, instead of showing the full inline form, replace with an edit button. That way users can simply skip all steps and their validations if the need is just changing the status of the Node or the display setting. `@TODO:` Add settings so this behavior can be enabled disabled but also add a 'Cancel Edit' button. that way if you press edit and the form displays you can go back to making it all dissapear without shame. Also: setting to display a 'preview' of the data that exists. Can be either via a template? but even better using the currently enabled Webform which includes a always a 'Display Preview' option. If we re-use that we have the best of two worlds. THIS LAST ONE NEEDS ITS OWN TICKET!