Allows WIKIdata elements to have only Label Visible
And still store and persist the URLs coming from WIKIDATA Universe.
This is a proof of concept and will be shifted mostlikely to a base class/abstract class that will be common to all our auth records!

This also deals with multiple elements, either each individually wrapped on rows of a table with headers. Not easy! But i like it!