For now. Just remove some of the excesive EXIF data from preview
This whole thing makes me not happy. Too much code in a preprocess stage we don't even use. So a, first make it a setting, move it out of the theme layer into the actual element and second, reuse as much as possible of the core SBF functionality. Pretty sure this responds to a super specific use case