Add a cancel edit button to any inline form that can be closed!
So this is new and was asked by the cool folks from ESIE (looks at @alliomeria :eyes: ). This pull adds dynamically a 'Cancel --some placeholder text you decide on' button when editing an existing object via a webform. Which has two effects.
- REmoves the current form and shows back the 'Edit  --some placeholder text you decide on' button shows back also the Save Node buttons on top.
If feels not fancy but brings also back the 'edit in a pop-up mode' and a more stronger and better code for all of this. That. Late. Last pull for webforms before official beta3. Adios