Humble attempt at making nominatim better
This is a response to a detail that bothers me:
When you search for a place and then select, the option you selected sticks during the whole session. Let's say selection has 4 results and you select number 4, if you search again, the selected option is again number 4, independently that this is another place in earth. Things get serious when the response has only one! In that case you can no longer select anything because its already selected. I need to double check if i will have to also modify the user input here or not