The Transplant element
Only transplant trees on months that don't contain an 'R' (applies to the southern hemisphere).

Really: This Webform element uses a Twig template to process input (other stored values in the webform submission, e.g a XML import) and generates (if the admin sets this correctly) a JSON encoded version of the new values.

Nice: it only replaces/fills up elements that EXIST in the FORM and are EMPTY. Means it will never replace a form manually edited by the user, nor values set by the user during Input.

TODO: add some config options to
a) trigger a full cleanup in the presence of a field change (like a new XML file was  uploaded) or this can be done in the XML file uploader ... need to chek
b) option to 'always replace'
c) suggest new value but keep the old one (could be done via a tooltip?)

I LOVE THIS ELEMENT!