Better WebformMetadata Upload Element
Different XML to JSON strategy with correct namespace calculation and also consistent always nested in Arrays values.

TODO: setting to allow some overrides, option to purge completely the import once the webform filling is done (or not) and hopefully an extra setting to allow the user to either use my original JsonSimpleXMLElementDecorator (slimmer but less consistent in data structure) or the new one, but also fix the number of files to 1 always and allow settings to be passed to the $simplexmltoarray class.