Fix multi valued and single valued elements now
This is a different approach but it seems to me a better option for searching/faceting and also twig templating:
Date values, if a single string will be split into pieces and classified into 3 possible categories, point date, range or free. Once that has happened, data will be saved with its classification and splitt. No validation yet here but then, each date element has its own. Only real validation i need is to make sure that in a range the to  date is later in time to the from date.