Webform changed the way it manages nesting for multiple compounds!
Gosh. This will be keep happening and i need to we watching closely.
This fixes an issue where `#vocab` is not bein saved when it matches the default. That is good because webforms don't carry values that can change in time inside their YAML, bad because we need that value from the element to react. THAT IS NEW! AND BAD. gosh gosh