Fix broken Autocomplete to node id resolving on Panorama Builder
This is also new to me, probably a newer version of Webform changed this? This should not be backwards breaking since we still apply the old strategy in case the passed value is wrong. And we do create an empty hotspot if all our checks fail user can remove if they want