Issue 60: New element: the transplant, XML upload and Date Range (#63)
* The Transplant element

Only transplant trees on months that don't contain an 'R' (applies to the southern hemisphere).

Really: This Webform element uses a Twig template to process input (other stored values in the webform submission, e.g a XML import) and generates (if the admin sets this correctly) a JSON encoded version of the new values.

Nice: it only replaces/fills up elements that EXIST in the FORM and are EMPTY. Means it will never replace a form manually edited by the user, nor values set by the user during Input.

TODO: add some config options to
a) trigger a full cleanup in the presence of a field change (like a new XML file was  uploaded) or this can be done in the XML file uploader ... need to chek
b) option to 'always replace'
c) suggest new value but keep the old one (could be done via a tooltip?)

I LOVE THIS ELEMENT!

* Better WebformMetadata Upload Element

Different XML to JSON strategy with correct namespace calculation and also consistent always nested in Arrays values.

TODO: setting to allow some overrides, option to purge completely the import once the webform filling is done (or not) and hopefully an extra setting to allow the user to either use my original JsonSimpleXMLElementDecorator (slimmer but less consistent in data structure) or the new one, but also fix the number of files to 1 always and allow settings to be passed to the $simplexmltoarray class.

* strawberryFieldharvester gets some bonus tracks

I enforce now the content entity -> originalValue property, even if in our case the entity is always new because (means i never have access to it) because we are NOT saving the submission. This helps with some custom webform elements that are coded to act on that property and our own code in the ComputedTransplant

* New Metadata date field

WIP, still not working properly. Form elements, states and inputs are a beast and i'm tired

* Add comment about prepareMultipleWrapper

* Gosh. It was always me. NOOO

* No longer parse the date here

Each individual subelement, date_from, date_to and date_free should/have/can do this. NO need to have a master parsing happening, makes just everythin more complex

* Fix multi valued and single valued elements now

This is a different approach but it seems to me a better option for searching/faceting and also twig templating:
Date values, if a single string will be split into pieces and classified into 3 possible categories, point date, range or free. Once that has happened, data will be saved with its classification and splitt. No validation yet here but then, each date element has its own. Only real validation i need is to make sure that in a range the to  date is later in time to the from date.

* Missing use

* Webform changed the way it manages nesting for multiple compounds!

Gosh. This will be keep happening and i need to we watching closely.
This fixes an issue where `#vocab` is not bein saved when it matches the default. That is good because webforms don't carry values that can change in time inside their YAML, bad because we need that value from the element to react. THAT IS NEW! AND BAD. gosh gosh

* Fix broken Autocomplete to node id resolving on Panorama Builder

This is also new to me, probably a newer version of Webform changed this? This should not be backwards breaking since we still apply the old strategy in case the passed value is wrong. And we do create an empty hotspot if all our checks fail user can remove if they want

* Don't override the main processWebformComposite if not present yet

* Adds and extra variation of the endpoint overwrite for multiple...

with table headers

* Override Drupal's Autocomplete Splitting by comma if SBF driven

This came from user/use case and it totally makes sense. Drupal splits autocomplete terms automatically by "," When filling our LoC or Wikidata elements it excludes all terms that have commas, which given how many of these terms are built leaves all narrow elements out.

This requires testing so will ask EADC @zenyzspalding and team to test and test over.

* VIAf for EADC

That, Now VIAF should work correctly. We get only the VIAF ID for now, but we should be able also show/fetch anything (biblissma)

* Update for webform 6.x

@giancarlobi this is for your Drupal 9.02 deployment

* Ups.

@giancarlobi committed too soon!

* Updates to XML and CSV import to allow subclassing

* CSV import element and some element generalizations

* Add D9 Webform 6 compatibilities

First move also to deprecate old ID given (by mistake!!) with an extra space two years ago. Makes theming,etc quite complex. An update hook will follow

* Allow free form to exist always.

@TODO need to be sure we clear when mode changes? Or do we always keep the free form ones?

* Be more strict with merging data

If data needs to be merged, then make sure it we are actually dealing with array v/s array

* remove debug statement in Date Element

* Make sure error message blames settings, not user interaction

This is in case the form is not correctly setup

* Fix Webform 6.0 Deprecation. Needs testing so will just push first

(FYI I have not tested this, this is just moving forward but not ready!)

* Issue 60 (#67)

* Syntax error introduced in parent committ

* Should fix summary and also replace old handler with new one!

@giancarlobi can you please test? The update hook needs to run

* Wrong formatting (only cosmetic)

* Adding a message at the end

In case we need some info?

* Famous base hook

Because hooks are such a great idea in 2020 ...

* So silly... how did i ended double nesting this theme?

Fixes Hook theme webform_handler .. summary not found
and actually shows the summary (in machine name language)

Co-authored-by: Giancarlo <giancarlo.birello@gmail.com>