Allow Standalon Webforms to be prepopulated using a ?do=UUID argument
This is fancy but not sure if it will be needed for many Archipelagos (right now self submission and workflows like OH are demanding this sacrifice) so we made it an optional in our SBF Harvested Handler. It does the following
- If there is a GET argument ?do=uuid in a stand alone webfor and/or a webform Node that uses a webform with an  SBF Harvester Handler and this option enabled and the UUID passed is of a ADO (a digital object) it will only once and on the first call get the values from the node (fear not permissions are checked) iterate over every WEbform form element filling up the values from the ADO, only and only if the user can create / submit those values. It also removes keys with ":". We can also in the future/next commit remove attached files if needed? (another option?)

I had to use an alter (sorry sorry) but its the only way to do this without a custom Controller and we have becuase of Oral Histories quite some diverse needs and links for this. Gosh