Adding actual saving of drafts on Webform navigation
@alliomeria we will have to talk about this...
So what i need now after all this is that we make sure people understand that when this is enabled either 1.- No saving is possible without validation. Right now on the navigation we are shamelessly bypassing validation.
Or
2.- This option is available only for admins?
or
3.- We only allow navigation that enforces at all pages to be validated. Which means, visiting each page, and not skipping any step

its a complex thing since a lot of the validation actually happens on a per page basis!

Tired now, so will keep developing tomorrow morning...