And le creme the la creme
Ok. Seem like progress saving is functional now!
- So much
- More than much
- ON wizard navigation we keep track of un handled errors and then we clear errors so people can jumo
- ON next/submit/preview button we check if the current step was validated and if so we clear the tracked errors (so people can actually correct them)
On any page that has a "submit/save" we check if there were errors tracked. If so we disable the submit and show a nice message.
TODO @alliomeria
: add the "quick save button" requested
Make sure that we also check that "EVERY" visible/accesible page was visited, not only the ones we found errors (tiny detail but is 1AM!)
TODO DEV too:
There is a lot of duplicated code in how we check if the node is new or not, if its widget driven, if it has wizard pages, etc. WE can do better and make a lot of that common methos

Of course this needs more work and a lot of testing, but right now i'm optimistic!