I tell people now directly how lazy they can be!
Yep. Do not wait over a week. Private gets cleaned after that if not renewed
I "could" allow the draft message to be set at the webform level BUT, since we are not submitting dear webform hides the settings UI making all that just a YML operation. We could add this message as a setting of the formatter OR of handler (better?)