Clean and simplify? (mmm) Metadata Date element
A few fixes and a lot of questions/doubts
- Dates can now be shown as RAW on the preview page
- When multiple empty elements are being automatically pushed into a list (because the admin or users decided so) we no longer end with empty entries in the JSON. We clean them up. basically anything that only has a date_type is perceived as emtpy and removed (preSave of the element which I SHOULD USE MORE!)

Restored the Labels for the 3 different text boxes but no solution yet for a JS that can on the fly on state change of the radios change the values from Point Date to Start Date.

I ahve to be honest with myself. This is a needed element and "fully functional" entry method right now, but i DO NOT like it. I feel the code needs a deep cleanup and a lot of testing. But we can not right now stop 2 weeks just to get this done in a perfect way. So i will roll with how it is. It works, its not perfect, its stable and gets the job done. Perfect, eventually. Clean, sooner than Perfect
@alliomeria