Drupal Coding Standards
@patdunlavey this is the part that was missing.
How to test?

Upload a simple CSV file (anything that has a few rows). Check if upload works and saving ends you with a JSON that contains a rows as JSON Objects.
Now open the CSV with a file editor and paste Windows characters inside (good test are those strange double quotes, save as NOT UTF8 (windows-iso characters), try to upload again, it should complain/or depending on the element settings, transform them into ?

I have tested this a lot so if you can't test no problem. I may add before RC3 also a Headers check (so we can limit what is uploaded) and also a row number limit (so JSON does not have to keep with 10,000 nested objects).

Thanks!