Europeana Entity Suggest Webform element and Controller + cache responses
Europeana Accepts types: agent, place, concept.

Requires: an api key set inside settings.php. For now Europeana (Alpha 0.5 V) has a single one for everyone but this can change.

To test put this:

$settings['webform_strawberryfield.europeana_entity_apikey'] = 'apidemo';
In your settings.php. If not the controller will add a message to the user (somewhere)

Also updates README.md AND!

Caches responses for a week! This should give some LoD endpoints some peace of mind. Caches are dependent on the user and 500 codes, 401 and 404 codes are not cached.