Try a better caching strategy. We could also go for the default ...
JSON cached response in the future. Only reason we go for this is to have more control but we may not need it

Also. If the call is internal (who calls == same as the server and there is no session we do bypass the 5 seconds wait per call)

`@TODO:` Too much of a fuzz to call URLs and Routes for internal calls. We may want to have public methods (not a controller?) for every LoD reconciliation endpoint for 1.1.x or 1.2.x